backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.7.4
pydantic-settings>=2.10.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
langchain>=1.2.0
langchain-core>=1.2.0
langchain-community>=0.4.0
faiss-cpu==1.7.4
openai>=1.3.0
pyyaml==6.0.1
